requirements.txt
# requirements.txt
python-telegram-bot==11.1.0
bs4
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
future==0.18.1
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage
googleapis-common-protos==1.6.0
grpcio==1.24.3
idna==2.8
lxml==4.4.1
matplotlib==3.1.1
Pillow==6.2.1
pprint==0.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
